서비스 만들기

+++ src/main/java/com/example/webproject/entity/Announcement.java
@@ -16,7 +16,7 @@ public class Announcement {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer noticeNumber;
+    private Long noticeNumber;
 
     @Column(nullable = false)
     private String noticeTitle;
