+++ src/main/java/com/example/webproject/repository/AnnouncementRepository.java
@@ -3,6 +3,6 @@ package com.example.webproject.repository;
 import com.example.webproject.entity.Announcement;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnnouncementRepository extends JpaRepository<Announcement, Integer> {
+public interface AnnouncementRepository extends JpaRepository<Announcement, Long> {
 
 }
