+++ src/main/java/com/example/webproject/repository/LoanRepository.java
@@ -3,6 +3,6 @@ package com.example.webproject.repository;
 import com.example.webproject.entity.Loan;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoanRepository extends JpaRepository<Loan, Integer> {
+public interface LoanRepository extends JpaRepository<Loan, Long> {
 
 }
